% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_to_excel.r
\name{add_additional_worksheet}
\alias{add_additional_worksheet}
\title{Adds data to Excel file}
\usage{
add_additional_worksheet(results, title, worksheet_name)
}
\arguments{
\item{results}{Results}

\item{title}{Title of new Excel file}

\item{worksheet_name}{Name of worksheet}
}
\description{
Takes results and adds it to existing Excel file. Only used after
\code{\link{save_to_excel}}
}
\author{
Sean Brunson
}
