% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.r
\name{adjust_limits}
\alias{adjust_limits}
\title{Function to adjust x and y axis of ggplot2 plot}
\usage{
adjust_limits(lower_bound, upper_bound, lower_bound_change = 0.1,
  upper_bound_change = 0.1)
}
\description{
Makes sure that limits are above or below highest and lowest value respectively
}
\author{
Sean Brunson
}
