% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting_functions.r
\name{as_comma_pdf}
\alias{as_comma_pdf}
\title{Changes data value to comma type for PDF}
\usage{
as_comma_pdf(x, digits = 2, format = "f", ...)
}
\description{
Changes data value to comma type for PDF
}
\author{
Sean Brunson
}
