% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting_functions.r
\name{as_money_pdf}
\alias{as_money_pdf}
\title{Changes data value to money type for PDF}
\usage{
as_money_pdf(x, digits = 2, format = "f", ...)
}
\description{
Changes data value to money type for PDF
}
\author{
Sean Brunson
}
