% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile.r
\name{compile_paper}
\alias{compile_paper}
\title{Compiles Paper}
\usage{
compile_paper(full = TRUE)
}
\arguments{
\item{full}{Creates full paper if true or just results if false}
}
\description{
Compiles index.Rmd file in main directory of project to write paper
}
\author{
Sean Brunson
}
