% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_paper_folders.r
\name{create_paper}
\alias{create_paper}
\title{Main Function to Setup Folder Structure}
\usage{
create_paper(paper_name)
}
\arguments{
\item{paper_name}{Name of paper PDF}
}
\description{
Creates folders within new paper project
}
\examples{
\dontrun{
paper_name <- "TestBook"

create_paper(paper_name)
}
}
\author{
Sean Brunson
}
