% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{inherit_pdf_document}
\alias{inherit_pdf_document}
\title{Sets up Latex PDF template}
\usage{
inherit_pdf_document(...)
}
\arguments{
\item{...}{Arguments passed to \code{pdf_document()}.}
}
\description{
Sets up Latex PDF template
}
\author{
Sean Brunson
}
