% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_standared_errors.r
\name{lm_cluster}
\alias{lm_cluster}
\title{Function to cluster standard errors for linear models}
\usage{
lm_cluster(data, formula, cluster)
}
\description{
Function to cluster standard errors for linear models
}
\examples{
\dontrun{
model <- lm_cluster(data=data, formula=y ~ x, cluster="cluster_id")

}
}
\author{
Sean Brunson
}
