% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{load_resources_if_missing}
\alias{load_resources_if_missing}
\title{Loads anything extra needed in the Latex template}
\usage{
load_resources_if_missing(template_name, resources)
}
\description{
Loads anything extra needed in the Latex template
}
\author{
Sean Brunson
}
