% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.r
\name{make_minimum_background}
\alias{make_minimum_background}
\title{Function to make plots as minimum as possible}
\usage{
make_minimum_background()
}
\description{
Function to make plots as minimum as possible
}
\author{
Sean Brunson
}
