% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.r
\name{palette_choice}
\alias{palette_choice}
\title{Setup color palette}
\usage{
palette_choice(palette = "three_color", reverse = FALSE, ...)
}
\description{
Setup color palette
}
\author{
Sean Brunson
}
