% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_to_excel.r
\name{remove_from_folder}
\alias{remove_from_folder}
\title{Removes old files from result's folder}
\usage{
remove_from_folder(title)
}
\arguments{
\item{title}{Title of old file}
}
\description{
Get all file names in result's folder, put in dataframe, and remove old ones
if name and date are old
}
\author{
Sean Brunson
}
