% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_to_excel.r
\name{save_to_excel}
\alias{save_to_excel}
\title{Saves data to Excel file}
\usage{
save_to_excel(results, title, worksheet_name = NULL)
}
\arguments{
\item{results}{Results}

\item{title}{Title of new Excel file}

\item{worksheet_name}{Name of worksheet}
}
\description{
Takes results and creates and saves to Excel file
}
\author{
Sean Brunson
}
