% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.r
\name{scale_fill_special}
\alias{scale_fill_special}
\title{Function to create own color fill columns}
\usage{
scale_fill_special(palette = "three_color", discrete = TRUE,
  reverse = FALSE, ...)
}
\arguments{
\item{palette}{A palette function that when called with a single integer
argument (the number of levels in the scale) returns the values that
they should take.}
}
\description{
Function to create own color fill columns
}
\author{
Sean Brunson
}
